requirements.txt
-f https://download.pytorch.org/whl/cu118/torch
-f https://download.pytorch.org/whl/cu118/torchvision
torch==2.2.1+cu118
torchvision==0.17.1+cu118
wandb==0.16.4
tqdm==4.66.2
pillow==10.2.0
matplotlib==3.8.3
numpy==1.26.3
scipy==1.13.0
torchview==0.2.6
graphviz==0.20.3
opencv-python==4.9.0.80
